Solved GoogleAPI call to be Global in app, added async map to Add via. global google.maps API, error with "this" within the scope of the map
