+++ src/GoogleMap/Gmap.js
@@ -1,64 +1,10 @@
 import React, { Component } from 'react';
-import regions from './geojson.json';
-import scriptLoader from 'react-async-script-loader'
 import { connect } from 'react-redux';
-import { testingMap } from '../actions/mapActions';
+
+
 const baseUrl = 'https://raw.githubusercontent.com/m-madden/lostandfound/master/';
 
-let google = window.google
-
-
-let dummyData = [{
-        Name: "Spot",
-        Color: "White",
-        Breed: "Great Dame",
-        Date: new Date().toString(),
-        Location: {
-          position: [{
-            lat: 36.5571445511326,
-            lng: -86.17903900146484
-          }],
-          key: this.name,
-          defaultAnimation: 2
-        },
-        Type: "Dog",
-        Status: ["Found"]
-      },{
-        Name: "Whiskers",
-        Color: "Orange",
-        Breed: "Evil",
-        Date: new Date().toString(),
-        Location: {
-          position: [{
-            lat: 36.0571445511326,
-            lng: -86.97903900146484
-          }],
-          key: this.name,
-          defaultAnimation: 2
-        },
-        Type: "Cat",
-        Status: ["Lost"]
-      },{
-        Name: "Rusty",
-        Color: "White with Black",
-        Breed: "Jack Russel",
-        Date: new Date().toString(),
-        Location: {
-          position: [{
-            lat: 36.9121445511326,
-            lng: -85.17903900146484
-          },
-          {
-            lat: 36.9121445511396,
-            lng: -85.77903900141484
-          }],
-          key: this.name,
-          defaultAnimation: 2
-        },
-        Type: "Dog",
-        Status: ["Lost", "Found"]
-      }
-    ]
+// This is the guide for all of it. 
 
 
 class Gmap extends Component {
@@ -67,35 +13,30 @@ class Gmap extends Component {
 
     this.state = {
       googleLoaded: false,
-      googleMap: {},
-      dummyData: dummyData
+      googleMap: {}
     }
   }
 
-  componentWillReceiveProps ({ isScriptLoaded, isScriptLoadSucceed }) {
-    if (isScriptLoaded && !this.props.isScriptLoaded) { // load finished
-      if (isScriptLoadSucceed) {
-        
-        let map = new google.maps.Map(this.refs.map, {
-          zoom: 16,
-          center: {
-            lat: 36.170295,
-            lng: -86.674846
-          }
-        });
+  componentDidUpdate() {
 
-        
-      // Comment out the three lines below to test functionality
-      // for(var i=0;i<regions.length;i++) {
-      //   map.data.add({geometry: new google.maps.Data.Polygon([regions[i].polygon])})
-      // }
+    let mapData = this.props.mapData
 
-      let markers = this.state.dummyData.map(animal => {
+    if (window.google.maps) {
+
+    let map = new window.google.maps.Map(document.getElementById("map"), {
+      zoom: 16,
+      center: {
+        lat: 36.170295,
+        lng: -86.674846
+      }
+    });
+
+    let markers = mapData.dummyData.map(animal => {
         animal.Location.position.map((position, index) => {
-          return new google.maps.Marker({
+          return new window.google.maps.Marker({
             position: position,
             map: map,
-            icon: pullIcon(animal, index)
+            // icon: pullIcon(animal, index)
           })
         })
       });
@@ -108,10 +49,10 @@ class Gmap extends Component {
 
       function findRegion(latLng) {
         var regionName;
-        for(var i=0; i<regions.length;i++) {
-          var currentPoly = new google.maps.Polygon({paths: regions[i].polygon});
-          if(google.maps.geometry.poly.containsLocation(latLng, currentPoly)) {
-            regionName = regions[i].name;
+        for(var i=0; i<mapData.geoJson.length;i++) {
+          var currentPoly = new window.google.maps.Polygon({paths: mapData.geoJson[i].polygon});
+          if(window.google.maps.geometry.poly.containsLocation(latLng, currentPoly)) {
+            regionName = mapData.geoJson[i].name;
           }
         }
         if(regionName) {
@@ -122,32 +63,34 @@ class Gmap extends Component {
       }
 
       function placeMarkerAndPanTo(latLng, map) {
-        var marker = new google.maps.Marker({
+        var marker = new window.google.maps.Marker({
           position: latLng,
           map,
         });
         map.panTo(latLng);
       }
+    }
+    
 
       function pullIcon(animal, index) {
         var status = animal.Status[index] === "Found" ? "found" : "lost";
         var type = animal.Type === "Cat" ? "Cat" : "Dog";
         return  baseUrl + status + type + "Icon.png"
       }
-    }
-  }
-  // else this.props.onError()
   }
   
   render() {
     return (
 
       <div className="mapContainer">
-        {}
-        <div ref="map" id="map"></div>
+        {window.google ? <div ref="map" id="map" style={{height: "100vh", width:"100vw"}}></div> : <h1>Loading...</h1>}
       </div>
     )
   }
 }
 
-export default Gmap;
+export default connect(state => {
+  return {
+    mapData: state.mapData
+  }
+})(Gmap);
