+++ src/GoogleMap/RegionGmap.js
@@ -43,4 +43,5 @@ class RegionGmap extends Component {
   }
 }
 
-export default scriptLoader(["https://maps.googleapis.com/maps/api/js?key=AIzaSyDiUupl6Z9qBY5J_IKupr44xM542C23Xiw&libraries=places,geometry"])(RegionGmap);
+// export default scriptLoader(["https://maps.googleapis.com/maps/api/js?key=AIzaSyDiUupl6Z9qBY5J_IKupr44xM542C23Xiw&libraries=places,geometry"])(RegionGmap);
+export default RegionGmap
