+++ src/actions/mapActions.js
@@ -2,9 +2,9 @@ export function testingMap(lat) {
   console.log(lat)
 }
 
-export function setGoogleMaps(googleMaps) {
+export function setGoogleMap(googleMap) {
   return {
     type: "SET_GOOGLE_MAPS",
-    payload: googleMaps
+    payload: googleMap
   }
 }
