+++ src/api/searchApi.js
@@ -13,13 +13,13 @@ module.exports = {
 		//Filter by searchText
 		filteredAnimals = filteredAnimals.filter((animal) => {
 			let name = animal.Name.toLowerCase();
-			let loc = animal.Location.toLowerCase();
+			// let loc = animal.Location.toLowerCase();
 			let breed = animal.Breed.toLowerCase();
 			let color = animal.Color.toLowerCase();
 
 			return (searchText.length === 0 || 
 				(name.indexOf(searchText.toLowerCase()) > -1) ||
-				(loc.indexOf(searchText.toLowerCase()) > -1) ||
+				// (loc.indexOf(searchText.toLowerCase()) > -1) ||
 				(breed.indexOf(searchText.toLowerCase()) > -1) ||
 				(color.indexOf(searchText.toLowerCase()) > -1)
 			);
