+++ src/reducers/index.js
@@ -3,7 +3,7 @@ import { combineReducers } from 'redux';
 import animal from './animalReducers';
 import animals from './animalsReducers';
 import animalsWithPics from './randomAnimalsReducers';
-import mapReducers from './mapReducers';
+import mapData from './mapReducers';
 import searchFields from './searchReducers';
 import user from './userReducers';
 
@@ -11,7 +11,7 @@ export default combineReducers({
   animal,
   animals,
   animalsWithPics,
-  mapReducers,
+  mapData,
   searchFields,
   user
 });
